# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是一個使用 Gin 框架構建的 CRUD API 服務，包含用戶管理、書籍管理、產品管理、文章管理、認證、限流器、工作池等功能。",
        "title": "Go Gin CRUD API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/accounts/balance": {
            "get": {
                "description": "獲取當前帳戶餘額",
                "tags": [
                    "account"
                ],
                "summary": "獲取帳戶餘額",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.AccountResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "設置帳戶初始餘額",
                "tags": [
                    "account"
                ],
                "summary": "設置餘額",
                "parameters": [
                    {
                        "description": "餘額",
                        "name": "balance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "balance": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.AccountResponse"
                        }
                    }
                }
            }
        },
        "/api/accounts/batch": {
            "post": {
                "description": "並發執行多筆交易（存款、取款、查詢）",
                "tags": [
                    "account"
                ],
                "summary": "批量執行交易",
                "parameters": [
                    {
                        "description": "交易操作列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.BatchTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.BatchTransactionResponse"
                        }
                    }
                }
            }
        },
        "/api/accounts/batch/random": {
            "post": {
                "description": "生成隨機交易操作並執行（用於演示併發）",
                "tags": [
                    "account"
                ],
                "summary": "執行隨機批量交易",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "交易數量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "是否添加延遲（用於演示）",
                        "name": "delay",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.BatchTransactionResponse"
                        }
                    }
                }
            }
        },
        "/api/accounts/deposit": {
            "post": {
                "description": "向帳戶存款",
                "tags": [
                    "account"
                ],
                "summary": "存款",
                "parameters": [
                    {
                        "description": "存款金額",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.TransactionResponse"
                        }
                    }
                }
            }
        },
...
        "/api/accounts/withdraw": {
            "post": {
                "description": "從帳戶取款",
                "tags": [
                    "account"
                ],
                "summary": "取款",
                "parameters": [
                    {
                        "description": "取款金額",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.TransactionResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-gin-crud_internal_dto.AccountResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                }
            }
        },
...
        "go-gin-crud_internal_dto.BatchTransactionRequest": {
            "type": "object",
            "required": [
                "operations"
            ],
            "properties": {
                "operations": {
                    "description": "-1: 取款, 0: 查詢, 1+: 存款",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "go-gin-crud_internal_dto.BatchTransactionResponse": {
            "type": "object",
            "properties": {
                "final_balance": {
                    "type": "integer"
                },
                "initial_balance": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-gin-crud_internal_dto.TransactionResponse"
                    }
                }
            }
        },
...
        "go-gin-crud_internal_dto.DepositRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "go-gin-crud_internal_dto.TransactionResponse": {
            "type": "object",
            "properties": {
                "after": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "before": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "operation": {
                    "description": "\"deposit\", \"withdraw\", \"balance\"",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "go-gin-crud_internal_dto.WithdrawRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "請在值前加上 \"Bearer \" 前綴，例如：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}